# src/openapi.yaml
openapi: 3.1.0
info:
  title: Vormex API
  description: |
    Professional social networking platform for students - LinkedIn + Instagram reels + GitHub profile + LeetCode stats style.
    
    ## Features
    - User authentication (Email/Password, Google OAuth)
    - GitHub OAuth integration with contribution calendar
    - Activity tracking with streaks (GitHub-style)
    - Professional profile fields (Skills, Experience, Education, Projects, Certificates, Achievements)
    - Unified content feed (Posts, Articles, Forum Q&A)
    - XP and leveling system
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
  version: 1.0.0
  contact:
    name: Vormex API Support
    email: support@vormex.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Current server origin
  - url: https://api.vormex.in/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
        details:
          type: string
          description: Additional error details
        code:
          type: string
          description: Error code (e.g., "TOKEN_EXPIRED", "NOT_CONNECTED")
...
paths:
...
  /upload/avatar:
    post:
      tags:
        - Upload
      summary: Upload profile picture
      description: |
        Upload profile picture (expects pre-cropped 1:1 aspect ratio from frontend).
        Backend resizes to 400x400 and converts to AVIF format.
        Max file size: 10 MB.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Pre-cropped image file (1:1 aspect ratio)
              required:
                - image
      responses:
        '200':
          description: Profile picture uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile picture uploaded successfully
                  avatarUrl:
                    type: string
                    example: https://vormex.b-cdn.net/profiles/avatars/user-123-1234567890.avif
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      name:
                        type: string
                      profileImage:
                        type: string
        '400':
          description: Invalid image or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Upload
      summary: Delete profile picture
      description: |
        Delete the current user's profile picture.
        Removes the image from Bunny.net storage and sets profileImage to null in the database.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile picture deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile picture deleted successfully
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      name:
                        type: string
                      profileImage:
                        type: string
                        nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /upload/banner:
    post:
      tags:
        - Upload
      summary: Upload banner image
      description: |
        Upload banner image (expects pre-cropped 3:1 aspect ratio from frontend).
        Backend resizes to 1500x500 and converts to AVIF format.
        Max file size: 10 MB.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Pre-cropped image file (3:1 aspect ratio)
              required:
                - image
      responses:
        '200':
          description: Banner uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Banner image uploaded successfully
                  bannerUrl:
                    type: string
                    example: https://vormex.b-cdn.net/profiles/banners/user-123-1234567890.avif
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      name:
                        type: string
                      bannerImageUrl:
                        type: string
        '400':
          description: Invalid image or file too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Upload
      summary: Delete banner image
      description: |
        Delete the current user's banner image.
        Removes the image from Bunny.net storage and sets bannerImageUrl to null in the database.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Banner image deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Banner image deleted successfully
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      name:
                        type: string
                      bannerImageUrl:
                        type: string
                        nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'